gsm.reporting: Good Statistical Monitoring Reporting

The Good Statistical Monitoring or 'gsm' suite of R packages provides a framework for statistical data monitoring. 'gsm.reporting' provides workflows to generate the reporting data model needed to generate reports.

See also

Author

Maintainer: Jeremy Wildfire jwildfire@gmail.com

Authors:

Laura Maxwell lkmaxwell23@gmail.com

Zelos Zhu zelos.zhu@atorusresearch.com

Other contributors:

Gilead Sciences [copyright holder]